requirements.txt
k2-connect
python-dotenv
requests 
flask
